Remove api.py facade; distribute entry points to their modules
api.py was a leftover public-facade from when __init__.py exported the API.
Its wrappers added only indirection:
- ask() orchestrator -> new query.py (read path, mirrors indexing.py write path)
- start_watcher() -> indexing.py (it drives ingestion)
- ingest_books() -> dropped; ingest_directory already defaults to BOOKS_DIR
- ingest_pdf() -> dropped; was an unused alias for ingest_source
cli.py imports the real modules directly.
