+++ src/rag/indexing.py
@@ -69,3 +69,42 @@ def ingest_directory(directory: Path | None = None) -> list[dict]:
             continue
         results.append(ingest_source(path))
     return results
+
+
+def start_watcher(books_dir: Path | None = None) -> None:
+    """
+    Start a blocking watchdog watcher on books_dir.
+    Performs a catch-up ingest on startup, then watches for new files live.
+    Not suitable for serverless deployment.
+    """
+    import time
+    from watchdog.observers import Observer
+    from watchdog.events import FileSystemEventHandler, FileCreatedEvent
+
+    books_dir = books_dir or BOOKS_DIR
+    books_dir.mkdir(parents=True, exist_ok=True)
+
+    class _Handler(FileSystemEventHandler):
+        def on_created(self, event: FileCreatedEvent) -> None:
+            path = Path(event.src_path)
+            if not event.is_directory and path.suffix.lower() in SUPPORTED_EXTENSIONS:
+                print(f"[watcher] New file detected: {path.name}")
+                try:
+                    result = ingest_source(path)
+                    print(f"[watcher] Ingested: {result['filename']}")
+                except Exception as e:
+                    print(f"[watcher] Ingestion failed for {path.name}: {e}")
+
+    print(f"[watcher] Scanning {books_dir} for unprocessed files...")
+    ingest_directory(books_dir)
+
+    observer = Observer()
+    observer.schedule(_Handler(), str(books_dir), recursive=False)
+    observer.start()
+    print(f"[watcher] Watching {books_dir}. Press Ctrl+C to stop.")
+    try:
+        while True:
+            time.sleep(1)
+    except KeyboardInterrupt:
+        observer.stop()
+    observer.join()
